Firewall: Add hook to support firewall
User can add some phone numbers to white list or black list
in the firewall provider. When new call comes, CallNotifier
will query the firewall provider to check whether the
new call is in the black list. If it is in the black list,
this call will be hung up automatically.

Change-Id: Ib1c61e4f28a78d5f4df0d6ac63297c9d8b788359