MSIM: Clean up the calls in DISCONNECTED properly.
When an active voice call is hungup at remote end and onDisconnect is
called in CallModeler, clean up the calls in DISCONNECTED state for the
subscription on which call is disconnected whether or not this subscription
is active subscription.

In current implementation clean up is done only
for active subscription, which results in call not moving to IDLE from
DISCONNECTED, if call disconnected at remote end is for LCH subscription.

Change-Id: Ia9adea56ed0211f3a3feee0763ae7ac0a37b807a
CRs-Fixed: 629435