Fix crash during back to back emergency calls.
Phone process crashes with RuntimeException while
unregister receiver because the receiver is not
registered.

Move code of register receiver at onCreate() to be
sure that receiver will register when call through
bindService() and startService() both the cases.

Add null checks to avoid null pointer exception in
onDestroy() of EmergencyCallbackModeService.

Change-Id: If73bffe251b9f8f64973bcac7399a903bc2183e6
CRs-Fixed: 641312