TeleService: Use the original connection to get incoming call subscription
In the case of firewall intercept the incoming call successfully, the connection
and call did not save into the mCallMap, so the function getCallFromMap() will
return null, and evently cause the NullPointerException.

Use the original connection to get incoming call subscription,
and delete the duration dialog when firewall intercepted the incoming call.

CRs-Fixed: 661178

Change-Id: I555eacb3bfc00961fe92376ce2179fe09e78a6dc