In change pin2 dialog, display enter puk2 if pin2 is blocked
The user is prompted to change the pin2 if the pin2 gets
blocked while editing a FDN contact. In this scenario, in the
current behavior when the user tries to change the pin2,
a prompt to enter the old pin2 is displayed; whereas, the
correct behavior is to prompt the user to enter puk2 code if
the pin2 is blocked.

When the FdnSetting UI is created, query the uicc card app for
the status of pin2. If the pin2 is blocked, reset the state of
the change pin2 dialog solely for puk2 use.

Change-Id: I8fca6be35ae2e8490ce36874afeef20537f57893
CRs-Fixed: 678940