MSIM: Fix for no 'New card available' prompt in multisim.
In multisim case, whenever a new card is inserted and mcc,
mnc config changes, the user is not shown the new card
availability prompt.This is due to SetSubscription activity
getting relaunched during config change, which dismisses the
'New card available' prompt.
Add mcc, mnc parameter for configChanges in manifest file
to fix the issue.

Change-Id: If63553d3fb52a00ee6965109444c469c33a553f7
CRs-Fixed: 686031