TeleService: Set Sub2 APN disabled when mobile data on sub1
When there is no SIM in sub2, the APN setting preference in mobile network
settings is gray. But when there is SIM in sub2 and mobile data on sub1,
the APN setting item is removed. The two conditions are not consistent.

Set Sub2 APN setting preference disabled when mobile data on sub1.

CRs-Fixed: 710183

Change-Id: I9478a25de2d2e3bf19bebb782da5ac791c37e1d2