TeleService: Start ringing after PHONE_NEW_RINGING_CONNECTION is handled done
The during of received the PHONE_NEW_RINGING_CONNECTION and PHONE_INCOMING_RING
event is very short. We have hanged up the incoming call in the blacklist when we
received the PHONE_NEW_RINGING_CONNECTION event, but we received PHONE_INCOMING_RING
event soon, and at this time the phone state was still RINGING, so we started ringing.

Add a flag to indicate if we have handled the PHONE_NEW_RINGING_CONNECTION event,
if so we start ringing.

CRs-Fixed: 711438

Change-Id: Ia43ea2928460b9914f8242fec6e7fcd9e556c4ab