TeleService: Change EmergencyDialer lauchMode to singleInstance
If PrimarySubSetting activity is created on the top of EmergencyDialer
activity in the task, this issue will occur. Because the flags used in
the intents to start EmergencyDialer is FLAG_ACTIVITY_NEW_TASK, so when
click emergency call button, it will not start a new activity, instead,
the task will be brought to the front of the screen with PrimarySubSetting
on top of EmergencyDialer, but PrimarySubSetting can not show when locked,
so there is no response and can not enter emergency call screen.

Change EmergencyDialer lauchMode to singleInstance, so PrimarySubSetting is
in different task from EmergencyDialer.

CRs-Fixed: 749624

Change-Id: I213fe58d27067e84405f5d16521d8ef9605d0668