Fix issue that no missing call notification shows after end the MT call.
The missed call cause depends on two values(notificationType/notificationCode).
And the issue happens because it receives the message UNSOL_SUPP_SVC_NOTIFICATION
during MO call, it leads to the disconnected call cause which became to restricted
after end the MO call.

As long as the message UNSOL_SUPP_SVC_NOTIFICATION is over, then reset the value
of the notificationtype/notificationCode.

CRs-Fixed: 759427

Change-Id: Iea5514b030a5c8fb82f84260ebfc4c4c055c54be