Fix IllegalStateException for disconnect cause codes.
The Util for mapping the telephony cause code to telecom cause
code is not complete and any new values added in telephony without
corresponding change in the util will through the exception. The
solution here is to map the unrecognized error code to cause
unknown intead of exception that leads to a crash.

Change-Id: Ibeda238764e196a68558f6955263ba775f87da3d
CRs-Fixed: 772015