IMS: NPE fix when call is migrated to a conference
The Supplementary Service notification events are not un-registered
properly whenever originalConnection is set during an IMS call.
This leads to multiple events generated causing null pointer exception
in the Handler that does not have the context of originalConnection.
This change unregisters all the required events that are registered
again when originalConnection is set.

Change-Id: I2fbf8bd1d9a139596d9bcd09ff64cba6da3f1c55
CRs-Fixed: 777609