TeleService: Do not create a PhoneAccount for the slot without SIM card
When the slot has no SIM card, the dummy subId now is no longer
negative number, so the TelecomAccountRegistry will create a PhoneAccount
for the slot without SIM card too.

Check getSubState() method to confirm whether the slot has SIM card or not,
only add a PhoneAccount for the slot with SIM card.

CRs-Fixed: 820101
Change-Id: I19934f735ab0d77a9f464ebb90d11574c8c16955