SEEK: Send correct responses for apdu transmit and simIO requests.
Correct the responses sent for IccTransmitApduLogicalChannel,
iccTransmitApduBasicChannel & iccExchangeSimIO requests:

When these requests fail with some response data from RIL,
send the response received from RIL to client.

But when the requests fail without any response data from RIL,
set the response to a a default value ( IccIoResult with sw1 = 0x6f,
sw2 = 0  and payload as null). Without this, the response for the client
is set to null and PhoneInterfaceManager enters an infinite loop while
processing the response.

Change-Id: I9494e17395a834b9a7c997c1afb5ef4b31b28e76