Fix to improve user experience when SIM PIN unlock fails.
Currently if SIM PIN is enabled and user enters the SIM PIN
when cross binding is in progress unlock fails. This scenario
cannot be avoided though as this is timing related.So the fix
is to let the user know there is no decrease in PIN retry attempts
by displaying appropriate toast.

Fix: In PhoneInterfaceManager, if SIM PIN request fails check for
cross mapping in progress and 'REQUEST_NOT_SUPPORTED' response and
map the result to cross mapping error value.

Change-Id: I2b809bc2342b435b785ba6cff2d7253ae7ad60df
CRs-Fixed: 854773.