Consider subscription deactivation from modem
When choosing subscription to place ecall, check if subscription
is deactivated from modem when sim state is ready. This is required
in a case where modem deactivates subscription for illegal sim event.
In such case sim state will come as ready as illegal is mapped to ready
and hence the need to check for sub deactivation from lower layers.

Change-Id: I275919eeab4565a9ef8f28220340cb094128582d
Fixed-CRs: 850188