Remove the back button on action bar
There is no back button on the action bar in portrait mode,
when we set setDisplayHomeAsUpEnabled fasle, but after
rotating DUT to landscape mode, preferenceActivity will be
created again then the back button is shown. So need to make
sure the back button will not be shown all the time.

CRs-Fixed: 993095

Change-Id: I6e75ae92b64637b160969c20f56d5e8fc63ff6ea