Dial Ecall in APM and No sim inserted
When no sim card is inserted and airlpane mode is enabled in the
device, Dial a Emergency call. When EmergencyCallHelper is trying to
place the call, radio is powered on. Due to no SIM is inserted, Modem
reports the service as Emergency only(limited service) but in telephony
Emergency only is also consired as OOS. Due to service state, emergemcy
call is placed only after all the retires are over which adds to the
delay in placing the call.

CRs-Fixed: 1006598
Change-Id: I8fdde2477c497c7460f9eaacb7f620f589959729